expand delimitor, double error message fixed, FINAL?

+++ srcs/execution/status_code.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   status_code.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lnjoh-tc <lnjoh-tc@student.42.fr>          +#+  +:+       +#+        */
+/*   By: jewu <jewu@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/17 14:36:11 by jewu              #+#    #+#             */
-/*   Updated: 2024/10/08 18:38:31 by lnjoh-tc         ###   ########.fr       */
+/*   Updated: 2024/10/11 15:18:19 by jewu             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,14 +15,26 @@
 extern volatile int	g_sig_flag;
 
 // Handle SIGQUIT signal in fork process
-void	handle_sigquit_in_fork(t_shell *gear_5, int cmd)
+void	handle_sig_in_fork(t_shell *gear_5, int cmd)
 {
 	int	i;
 
+	if (!gear_5)
+		return ;
 	i = -1;
-	while (++i < cmd)
-		kill(gear_5->pid_tab[i], SIGQUIT);
-	gear_5->exit_status = 131;
+	if (g_sig_flag == SIGQUIT)
+	{
+		while (++i < cmd)
+			kill(gear_5->pid_tab[i], SIGQUIT);
+		gear_5->exit_status = 131;
+	}
+	else if (g_sig_flag == SIGINT)
+	{
+		while (++i < cmd)
+			kill(gear_5->pid_tab[i], SIGINT);
+		gear_5->exit_status = 130;
+	}
+	g_sig_flag = 0;
 }
 
 //update parent exit status with last child status code
@@ -43,12 +55,18 @@ int	child_status_code(t_shell *gear_5)
 		exit = 127;
 	else if (gear_5->status_code == 0)
 		exit = 0;
-	else if (gear_5->status_code == 3)
+	else if (gear_5->status_code == 131)
 		exit = 131;
+	else if (gear_5->status_code == 130)
+		exit = 130;
+	else if (gear_5->status_code)
+		exit = gear_5->status_code;
 	return (exit);
 }
 
 //get status code of the last child process
+//WIFEXITED = when child process ended normally
+//WIFSIGNALED = when a signal has been sent to child
 int	get_status_code(t_shell *gear_5, int cmd)
 {
 	int	i;
@@ -66,11 +84,10 @@ int	get_status_code(t_shell *gear_5, int cmd)
 			if (WIFEXITED(status))
 				last_status = WEXITSTATUS(status);
 			else if (WIFSIGNALED(status))
-				last_status = WTERMSIG(status);
+				last_status = 128 + WTERMSIG(status);
 		}
 	}
 	g_sig_flag = 0;
-	handle_signal();
 	return (last_status);
 }
